backend/requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
langchain==1.2.15
openai==2.30.0
faiss-cpu==1.13.2
pydantic==2.12.5
python-dotenv==1.2.2
